+++ Xenity_Engine/Source/editor/command/commands/inspector_commands.h
@@ -449,8 +449,8 @@ inline void InspectorCreateGameObjectCommand::Undo()
 		}
 		else
 		{
-			const int targetCount = targets.size();
-			for (int i = 0; i < targetCount; i++)
+			const size_t targetCount = targets.size();
+			for (size_t i = 0; i < targetCount; i++)
 			{
 				if (oldParents[i].lock())
 					targets[i].lock()->SetParent(oldParents[i].lock());
